+++ src/utils.ts
@@ -1,193 +1,8 @@
 import chalk from 'chalk';
 import console_stamp from 'console-stamp';
 import * as puppeteer from 'puppeteer-core';
-import { scrollPageToBottom } from 'puppeteer-autoscroll-down';
-import * as fs from 'fs-extra';
 
 console_stamp(console);
-
-let contentHTML = '';
-export interface GeneratePDFOptions {
-  initialDocURLs: Array<string>;
-  excludeURLs: Array<string>;
-  outputPDFFilename: string;
-  pdfMargin: puppeteer.PDFOptions['margin'];
-  contentSelector: string;
-  paginationSelector: string;
-  // deprecated - user paperFormat
-  pdfFormat?: puppeteer.PaperFormat;
-  paperFormat: puppeteer.PaperFormat;
-  excludeSelectors: Array<string>;
-  cssStyle: string;
-  puppeteerArgs: Array<string>;
-  coverTitle: string;
-  coverImage: string;
-  disableTOC: boolean;
-  coverSub: string;
-  waitForRender: number;
-  headerTemplate: string;
-  footerTemplate: string;
-  protocolTimeout: number;
-  filterKeyword: string;
-}
-
-/* c8 ignore start */
-export async function generatePDF({
-  initialDocURLs,
-  excludeURLs,
-  outputPDFFilename = 'docs-to-pdf.pdf',
-  pdfMargin = { top: 32, right: 32, bottom: 32, left: 32 },
-  contentSelector,
-  paginationSelector,
-  paperFormat,
-  excludeSelectors,
-  cssStyle,
-  puppeteerArgs,
-  coverTitle,
-  coverImage,
-  disableTOC,
-  coverSub,
-  waitForRender,
-  headerTemplate,
-  footerTemplate,
-  protocolTimeout,
-  filterKeyword,
-}: GeneratePDFOptions): Promise<void> {
-  const execPath =
-    process.env.PUPPETEER_EXECUTABLE_PATH ?? puppeteer.executablePath('chrome');
-  console.debug(chalk.cyan(`Using Chromium from ${execPath}`));
-  const browser = await puppeteer.launch({
-    headless: 'new',
-    executablePath: execPath,
-    args: puppeteerArgs,
-    protocolTimeout: protocolTimeout,
-  });
-
-  const chromeTmpDataDir = browser
-    .process()
-    ?.spawnargs.find((arg) => arg.startsWith('--user-data-dir'))
-    ?.split('=')[1] as string;
-  console.debug(chalk.cyan(`Chrome user data dir: ${chromeTmpDataDir}`));
-
-  const page = await browser.newPage();
-
-  // Block PDFs as puppeteer can not access them
-  await page.setRequestInterception(true);
-  page.on('request', (request) => {
-    if (request.url().endsWith('.pdf')) {
-      console.log(chalk.yellowBright(`ignore pdf: ${request.url()}`));
-      request.abort();
-    } else request.continue();
-  });
-
-  for (const url of initialDocURLs) {
-    let nextPageURL = url;
-
-    // Create a list of HTML for the content section of all pages by looping
-    while (nextPageURL) {
-      console.log(chalk.cyan(`Retrieving html from ${nextPageURL}`));
-
-      // Go to the page specified by nextPageURL
-      await page.goto(`${nextPageURL}`, {
-        waitUntil: 'networkidle0',
-        timeout: 0,
-      });
-      if (waitForRender) {
-        console.log(chalk.green('Waiting for render...'));
-        await new Promise((r) => setTimeout(r, waitForRender));
-      }
-
-      // Make joined content html
-      if (excludeURLs && excludeURLs.includes(nextPageURL)) {
-        console.log(chalk.green('This URL is excluded.'));
-      } else if (filterKeyword && !(await matchKeyword(page, filterKeyword))) {
-        console.log(
-          chalk.yellowBright(
-            `Page excluded by keyword filter: ${filterKeyword}`,
-          ),
-        );
-      } else {
-        // Get the HTML string of the content section.
-        contentHTML += await getHtmlContent(page, contentSelector);
-        console.log(chalk.green('Success'));
-      }
-
-      // Find next page url before DOM operations
-      nextPageURL = await findNextUrl(page, paginationSelector);
-    }
-  }
-
-  console.log(chalk.cyan('Start generating PDF...'));
-
-  // Generate cover Image if declared
-  let coverImageHtml = '';
-  if (coverImage) {
-    console.log(chalk.cyan('Get coverImage...'));
-    const image = await getCoverImage(page, coverImage);
-    coverImageHtml = generateImageHtml(image.base64, image.type);
-  }
-
-  // Generate Cover
-  console.log(chalk.cyan('Generate cover...'));
-  const coverHTML = generateCoverHtml(coverTitle, coverImageHtml, coverSub);
-
-  // Generate Toc
-  const { modifiedContentHTML, tocHTML } = generateToc(contentHTML);
-
-  // Restructuring the HTML of a document
-  console.log(chalk.cyan('Restructuring the html of a document...'));
-
-  // Go to initial page
-  await page.goto(`${initialDocURLs[0]}`, { waitUntil: 'networkidle0' });
-
-  await page.evaluate(
-    concatHtml,
-    coverHTML,
-    tocHTML,
-    modifiedContentHTML,
-    disableTOC,
-  );
-
-  // Remove unnecessary HTML by using excludeSelectors
-  if (excludeSelectors) {
-    console.log(chalk.cyan('Remove unnecessary HTML...'));
-    await removeExcludeSelector(page, excludeSelectors);
-  }
-
-  // Add CSS to HTML
-  if (cssStyle) {
-    console.log(chalk.cyan('Add CSS to HTML...'));
-    await page.addStyleTag({ content: cssStyle });
-  }
-
-  // Scroll to the bottom of the page with puppeteer-autoscroll-down
-  // This forces lazy-loading images to load
-  console.log(chalk.cyan('Scroll to the bottom of the page...'));
-  await scrollPageToBottom(page, {}); //cast to puppeteer-core type
-
-  // Generate PDF
-  console.log(chalk.cyan('Generate PDF...'));
-  await page.pdf({
-    path: outputPDFFilename,
-    format: paperFormat,
-    printBackground: true,
-    margin: pdfMargin,
-    displayHeaderFooter: !!(headerTemplate || footerTemplate),
-    headerTemplate,
-    footerTemplate,
-    timeout: 0,
-  });
-
-  console.log(chalk.green(`PDF generated at ${outputPDFFilename}`));
-  await browser.close();
-  console.log(chalk.green('Browser closed'));
-
-  if (chromeTmpDataDir !== null) {
-    fs.removeSync(chromeTmpDataDir);
-  }
-  console.debug(chalk.cyan('Chrome user data dir removed'));
-}
-/* c8 ignore stop */
 
 /**
  * Checks whether a page contains a given keyword
@@ -201,8 +16,20 @@ export async function matchKeyword(page: puppeteer.Page, keyword: string) {
       "head > meta[name='keywords']",
       (element) => element.content,
     );
-    return metaKeywords.split(',').includes(keyword);
-  } catch {
+    if (metaKeywords.split(',').includes(keyword)) {
+      console.log(
+        chalk.green('Keyword found: ' + keyword + ' in ' + metaKeywords),
+      );
+      return true;
+    }
+    console.log(
+      chalk.yellowBright(
+        'Keyword not found: ' + keyword + ' in ' + metaKeywords,
+      ),
+    );
+    return false;
+  } catch (e) {
+    console.error(chalk.red('No meta keywords found: ' + e));
     return false;
   }
 }
@@ -220,7 +47,7 @@ export async function getHtmlContent(page: puppeteer.Page, selector: string) {
 
 /**
  * Retrieves the HTML content of an element matching the specified selector.
- * Adds page break styling and opens <details> tags for PDF generation.
+ * Adds page break styling for PDF generation.
  *
  * @param selector - The CSS selector of the element to retrieve.
  * @returns The HTML content of the matched element, or an empty string if no element is found.
@@ -231,15 +58,46 @@ export function getHtmlFromSelector(selector: string): string {
     // Add pageBreak for PDF
     element.style.pageBreakAfter = 'always';
 
-    // Open <details> tag
-    const detailsArray = element.getElementsByTagName('details');
-    Array.from(detailsArray).forEach((element) => {
-      element.open = true;
-    });
-
     return element.outerHTML;
   } else {
     return '';
+  }
+}
+
+type ClickFunction = (element: puppeteer.ElementHandle) => Promise<void>;
+type WaitFunction = (milliseconds: number) => Promise<void>;
+
+/**
+ * Recursively opens all <details> elements on a page.
+ *
+ * @param page - The Puppeteer page instance.
+ * @param clickFunction - A function to click the summary element of a <details> element.
+ * @param waitFunction - A function to wait for a specified number of milliseconds.
+ */
+export async function openDetails(
+  page: puppeteer.Page,
+  clickFunction?: ClickFunction,
+  waitFunction?: WaitFunction,
+) {
+  const detailsHandles = await page.$$('details');
+
+  console.debug(`Found ${detailsHandles.length} elements`);
+
+  for (const detailsHandle of detailsHandles) {
+    const summaryHandle = await detailsHandle.$('summary');
+    if (summaryHandle) {
+      console.debug(
+        `Clicking summary: ${await summaryHandle.evaluate(
+          (node) => node.textContent,
+        )}`,
+      );
+      await (clickFunction
+        ? clickFunction(summaryHandle)
+        : summaryHandle.click());
+      await (waitFunction
+        ? waitFunction(800)
+        : new Promise((r) => setTimeout(r, 800)));
+    }
   }
 }
 
@@ -283,10 +141,17 @@ export function concatHtml(
   toc: string,
   content: string,
   disable: boolean,
+  baseUrl: string,
 ) {
   // Clear the body content
   const body = document.body;
   body.innerHTML = '';
+
+  // Add base tag for relative links
+  // see: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/base
+  if (baseUrl) {
+    body.innerHTML += `<base href="${baseUrl}" />`;
+  }
 
   // Add the cover HTML to the body
   body.innerHTML += cover;
@@ -511,4 +376,53 @@ export function removeElementFromSelector(selector: string): void {
 
   // Remove each matched element
   matches.forEach((match) => match.remove());
+}
+
+/**
+ * Check if a page should be kept based on exclusion and filtering conditions.
+ *
+ * @param page - The Puppeteer page instance
+ * @param nextPageURL The URL of the next page to be checked.
+ * @param urlPath The path of the URL to be checked.
+ * @param excludeURLs List of URLs to exclude.
+ * @param filterKeyword Keyword to filter pages.
+ * @param excludePaths List of path patterns to exclude.
+ * @param restrictPaths Whether to restrict by path.
+ * @returns True if the page should be kept, false if it should be excluded.
+ */
+export async function isPageKept(
+  page: puppeteer.Page,
+  nextPageURL: string,
+  urlPath: string,
+  excludeURLs: string[] | undefined,
+  filterKeyword: string | undefined,
+  excludePaths: string[] | undefined,
+  restrictPaths: boolean | undefined,
+): Promise<boolean> {
+  if (excludeURLs && excludeURLs.includes(nextPageURL)) {
+    console.log(chalk.green('This URL is excluded.'));
+    return false;
+  } else if (filterKeyword && !(await matchKeyword(page, filterKeyword))) {
+    console.log(
+      chalk.yellowBright(`Page excluded by keyword filter: ${filterKeyword}`),
+    );
+    return false;
+  } else if (
+    excludePaths &&
+    excludePaths.some((path) => nextPageURL.includes(path))
+  ) {
+    console.log(
+      chalk.yellowBright(`Page excluded by path filter: ${excludePaths}`),
+    );
+    return false;
+  } else if (restrictPaths && nextPageURL.includes(urlPath) === false) {
+    console.log(
+      chalk.yellowBright(
+        `Page excluded by path restriction: ${urlPath} !== ${urlPath}`,
+      ),
+    );
+    return false;
+  }
+
+  return true;
 }